requirements.txt
pyTelegramBotAPI
requests
pandas
matplotlib
apscheduler
psycopg2-binary
